prediction/requirements.txt
Flask
Pillow
Torchvision